requirements.txt
streamlit
altair
numpy
pandas
#time
Pillow
matplotlib
#base64
requests
#dotenv
#os
# opencv-python
subprocess
